Compile unit split for deviceconfig and webserver
