+++ src/deviceconfig.cpp
@@ -40,66 +40,12 @@
 
 #include "audioservice.h"
 #include "deviceconfig.h"
+#include "deviceconfig_internal.h"
 #include "effectmanager.h"
 #include "jsonserializer.h"
 #include "systemcontainer.h"
 
 extern const char timezones_start[] asm("_binary_config_timezones_json_start");
-
-namespace
-{
-    constexpr const char* kRecompileNeededMessage = "recompile needed";
-
-    // It's annoying to have to map from the compile-time COLOR_ORDER macro to the
-    // runtime enum, but it is what it is.  It's better than tying to a FastLED type.
-
-    constexpr DeviceConfig::WS281xColorOrder ToRuntimeColorOrder(EOrder order)
-    {
-        switch (order)
-        {
-            case EOrder::RGB: return DeviceConfig::WS281xColorOrder::RGB;
-            case EOrder::RBG: return DeviceConfig::WS281xColorOrder::RBG;
-            case EOrder::GRB: return DeviceConfig::WS281xColorOrder::GRB;
-            case EOrder::GBR: return DeviceConfig::WS281xColorOrder::GBR;
-            case EOrder::BRG: return DeviceConfig::WS281xColorOrder::BRG;
-            case EOrder::BGR: return DeviceConfig::WS281xColorOrder::BGR;
-            default:          return DeviceConfig::WS281xColorOrder::GRB;
-        }
-    }
-
-    #if USE_WS281X
-    constexpr auto kCompiledWS281xColorOrder = ToRuntimeColorOrder(COLOR_ORDER);
-    #else
-    constexpr auto kCompiledWS281xColorOrder = DeviceConfig::WS281xColorOrder::GRB;
-    #endif
-
-    constexpr std::array<int8_t, NUM_CHANNELS> kCompiledWS281xPins = {
-        #if NUM_CHANNELS >= 1
-        LED_PIN0,
-        #endif
-        #if NUM_CHANNELS >= 2
-        LED_PIN1,
-        #endif
-        #if NUM_CHANNELS >= 3
-        LED_PIN2,
-        #endif
-        #if NUM_CHANNELS >= 4
-        LED_PIN3,
-        #endif
-        #if NUM_CHANNELS >= 5
-        LED_PIN4,
-        #endif
-        #if NUM_CHANNELS >= 6
-        LED_PIN5,
-        #endif
-        #if NUM_CHANNELS >= 7
-        LED_PIN6,
-        #endif
-        #if NUM_CHANNELS >= 8
-        LED_PIN7,
-        #endif
-    };
-}
 
 // DeviceConfig holds, persists and loads device-wide configuration settings. Effect-specific settings should
 // be managed using overrides of the respective methods in LEDStripEffect (mainly FillSettingSpecs(),
@@ -127,7 +73,7 @@ void DeviceConfig::SaveToJSON() const
 
 std::array<int8_t, NUM_CHANNELS> DeviceConfig::GetCompiledWS281xPins()
 {
-    return kCompiledWS281xPins;
+    return DeviceConfigInternal::GetCompiledWS281xPins();
 }
 
 const char* DeviceConfig::DriverName(OutputDriver driver)
@@ -145,7 +91,7 @@ const char* DeviceConfig::DriverName(OutputDriver driver)
 
 DeviceConfig::WS281xColorOrder DeviceConfig::GetCompiledWS281xColorOrder()
 {
-    return kCompiledWS281xColorOrder;
+    return DeviceConfigInternal::GetCompiledWS281xColorOrder();
 }
 
 String DeviceConfig::GetColorOrderName(WS281xColorOrder colorOrder)
@@ -160,80 +106,6 @@ String DeviceConfig::GetColorOrderName(WS281xColorOrder colorOrder)
         case WS281xColorOrder::BGR: return "BGR";
         default:                    return "GRB";
     }
-}
-
-const char* DeviceConfig::OutputDriverName(OutputDriver driver)
-{
-    return DriverName(driver);
-}
-
-const char* DeviceConfig::WS281xColorOrderName(WS281xColorOrder colorOrder)
-{
-    switch (colorOrder)
-    {
-        case WS281xColorOrder::RGB: return "RGB";
-        case WS281xColorOrder::RBG: return "RBG";
-        case WS281xColorOrder::GRB: return "GRB";
-        case WS281xColorOrder::GBR: return "GBR";
-        case WS281xColorOrder::BRG: return "BRG";
-        case WS281xColorOrder::BGR: return "BGR";
-        default:                    return "GRB";
-    }
-}
-
-std::optional<DeviceConfig::OutputDriver> DeviceConfig::ParseOutputDriverName(const String& name)
-{
-    if (name == "hub75")
-        return OutputDriver::HUB75;
-    if (name == "ws281x")
-        return OutputDriver::WS281x;
-
-    return std::nullopt;
-}
-
-std::optional<DeviceConfig::WS281xColorOrder> DeviceConfig::ParseWS281xColorOrderName(const String& name)
-{
-    if (name == "RGB") return WS281xColorOrder::RGB;
-    if (name == "RBG") return WS281xColorOrder::RBG;
-    if (name == "GRB") return WS281xColorOrder::GRB;
-    if (name == "GBR") return WS281xColorOrder::GBR;
-    if (name == "BRG") return WS281xColorOrder::BRG;
-    if (name == "BGR") return WS281xColorOrder::BGR;
-
-    return std::nullopt;
-}
-
-void DeviceConfig::AppendPins(JsonArray target, const std::array<int8_t, NUM_CHANNELS>& pins)
-{
-    for (auto pin : pins)
-        target.add(pin);
-}
-
-std::optional<int> DeviceConfig::ResolveUnifiedAudioInputPin(JsonObjectConst device, String* errorMessage)
-{
-    std::optional<int> requestedAudioInputPin;
-
-    if (device[DeviceConfig::AudioInputPinTag].is<int>())
-        requestedAudioInputPin = device[DeviceConfig::AudioInputPinTag].as<int>();
-
-    if (device["audio"].is<JsonObjectConst>())
-    {
-        auto audio = device["audio"].as<JsonObjectConst>();
-        if (audio["audioInputPin"].is<int>())
-        {
-            const int nestedAudioInputPin = audio["audioInputPin"].as<int>();
-            if (requestedAudioInputPin.has_value() && requestedAudioInputPin.value() != nestedAudioInputPin)
-            {
-                if (errorMessage)
-                    *errorMessage = "Malformed request";
-                return std::nullopt;
-            }
-
-            requestedAudioInputPin = nestedAudioInputPin;
-        }
-    }
-
-    return requestedAudioInputPin;
 }
 
 bool DeviceConfig::IsHub75Build()
@@ -422,8 +294,8 @@ bool DeviceConfig::DeserializeFromJSON(const JsonObjectConst& jsonObject, bool s
         }
     }
 
-    String runtimeConfigError;
-    if (!SetRuntimeConfig(updated, true, &runtimeConfigError))
+    auto [runtimeConfigValid, runtimeConfigError] = SetRuntimeConfig(updated, true);
+    if (!runtimeConfigValid)
         debugW("Ignoring invalid persisted runtime config: %s", runtimeConfigError.c_str());
 
     if (ntpServer.isEmpty())
@@ -441,326 +313,6 @@ bool DeviceConfig::DeserializeFromJSON(const JsonObjectConst& jsonObject, bool s
 void DeviceConfig::RemovePersisted()
 {
     RemoveJSONFile(DEVICE_CONFIG_FILE);
-}
-
-const std::vector<std::reference_wrapper<SettingSpec>>& DeviceConfig::GetSettingSpecs()
-{
-    if (settingSpecs.empty())
-    {
-        // Section keys used below correspond to entries in the section catalog
-        // emitted by FillUnifiedSettingsSchemaJson(). Keep them in sync.
-        constexpr const char* kSectionSystem     = "system";
-        constexpr const char* kSectionLocation   = "location";
-        constexpr const char* kSectionClock      = "clock";
-        constexpr const char* kSectionAudio      = "audio";
-        constexpr const char* kSectionAppearance = "appearance";
-        constexpr const char* kSectionTopology   = "topology";
-        constexpr const char* kSectionOutput     = "output";
-
-        // ---- system section ------------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name           = HostnameTag,
-            .FriendlyName   = "Hostname",
-            .Description    = "The hostname of the device. A reboot is required after changing this.",
-            .Type           = SettingSpec::SettingType::String,
-            .EmptyAllowed   = true,
-            .Section        = kSectionSystem,
-            .RequiresReboot = true,
-            .ApiPath        = "device.hostname"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name          = PowerLimitTag,
-            .FriendlyName  = "Power limit",
-            .Description   = "The maximum power in mW that the matrix attached to the board is allowed to use. As the previous sentence implies, this "
-                             "setting only applies if a matrix is used.",
-            .Type          = SettingSpec::SettingType::Integer,
-            .HasValidation = true,
-            .MinimumValue  = (double)POWER_LIMIT_MIN,
-            .Section       = kSectionSystem,
-            .ApiPath       = "device.powerLimit"
-        }));
-
-        // ---- location section ----------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = LocationTag,
-            .FriendlyName = "Location",
-            .Description  = "The location (city or postal code) where the device is located.",
-            .Type         = SettingSpec::SettingType::String,
-            .Section      = kSectionLocation,
-            .ApiPath      = "device.location"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = LocationIsZipTag,
-            .FriendlyName = "Location is postal code",
-            .Description  = "Indicates if the value for the \"Location\" setting is a postal code (yes if checked) or not.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionLocation,
-            .ApiPath      = "device.locationIsZip"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = CountryCodeTag,
-            .FriendlyName = "Country code",
-            .Description  = "The <a href=\"https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2\">ISO 3166-1 alpha-2</a> country "
-                            "code for the country that the device is located in.",
-            .Type         = SettingSpec::SettingType::String,
-            .Section      = kSectionLocation,
-            .ApiPath      = "device.countryCode",
-            .Widget       = SettingSpec::WidgetKind::Select,
-            .Options      = SettingSpec::OptionsSource::IntlCountryCodes
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name               = TimeZoneTag,
-            .FriendlyName       = "Time zone",
-            .Description        = "The timezone the device resides in, in <a href=\"https://en.wikipedia.org/wiki/Tz_database\">tz database</a> format. "
-                                  "The list of available timezone identifiers can be found in the <a href=\"/timezones.json\">timezones.json</a> file.",
-            .Type               = SettingSpec::SettingType::String,
-            .Section            = kSectionLocation,
-            .ApiPath            = "device.timeZone",
-            .Widget             = SettingSpec::WidgetKind::Select,
-            .Options            = SettingSpec::OptionsSource::ExternalTimeZones,
-            .OptionsExternalUrl = "/timezones.json"
-        }));
-
-        // ---- clock section -------------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = Use24HourClockTag,
-            .FriendlyName = "Use 24 hour clock",
-            .Description  = "Indicates if time should be shown in 24-hour format (yes if checked) or 12-hour AM/PM format.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionClock,
-            .ApiPath      = "device.use24HourClock"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = UseCelsiusTag,
-            .FriendlyName = "Use degrees Celsius",
-            .Description  = "Indicates if temperatures should be shown in degrees Celsius (yes if checked) or degrees Fahrenheit.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionClock,
-            .ApiPath      = "device.useCelsius"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = NTPServerTag,
-            .FriendlyName = "NTP server address",
-            .Description  = "The hostname or IP address of the NTP server to be used for time synchronization.",
-            .Type         = SettingSpec::SettingType::String,
-            .Section      = kSectionClock,
-            .ApiPath      = "device.ntpServer"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name          = OpenWeatherApiKeyTag,
-            .FriendlyName  = "Open Weather API key",
-            .Description   = "The API key for the <a href=\"https://openweathermap.org/api\">Weather API provided by Open Weather Map</a>.",
-            .Type          = SettingSpec::SettingType::String,
-            .HasValidation = true,
-            .Access        = SettingSpec::SettingAccess::WriteOnly,
-            .Section       = kSectionClock,
-            .ApiPath       = "device.openWeatherApiKey"
-        }));
-
-        // ---- audio section -------------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name           = AudioInputPinTag,
-            .FriendlyName   = "Audio input pin",
-            .Description    = "External microphone input pin. This is boot-applied today because the audio task still owns the active DMA/I2S handles once sampling starts.",
-            .Type           = SettingSpec::SettingType::Integer,
-            .HasValidation  = true,
-            .MinimumValue   = -1.0,
-            .MaximumValue   = 48.0,
-            .Section        = kSectionAudio,
-            .RequiresReboot = !SupportsLiveAudioInputReconfigure(),
-            .ApiPath        = "device.audioInputPin"
-        }));
-
-        // ---- appearance section --------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name          = BrightnessTag,
-            .FriendlyName  = "Brightness",
-            .Description   = "Overall brightness the connected LEDs or matrix should be run at.",
-            .Type          = SettingSpec::SettingType::Integer,
-            .HasValidation = true,
-            .Section       = kSectionAppearance,
-            .ApiPath       = "device.brightness",
-            .Widget        = SettingSpec::WidgetKind::Slider,
-            // Display as 5..100 percent over the raw 13..255 range. The endpoints
-            // mirror what the legacy UI did: clamp at 5% on the low side, 100% at
-            // raw max, with linear remapping in between. The UI does the math.
-            .DisplayRawMin = (double)BRIGHTNESS_MIN,
-            .DisplayRawMax = (double)BRIGHTNESS_MAX,
-            .DisplayMin    = 5.0,
-            .DisplayMax    = 100.0,
-            .DisplaySuffix = "%"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = GlobalColorTag,
-            .FriendlyName = "Global color",
-            .Description  = "Main color that is applied to all those effects that support using it.",
-            .Type         = SettingSpec::SettingType::Color,
-            .Section      = kSectionAppearance,
-            .ApiPath      = "device.globalColor"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = SecondColorTag,
-            .FriendlyName = "Second color",
-            .Description  = "Second color that is used to create a global palette in combination with the current global color. That palette is used "
-                            "by some effects. Defaults to the <em>previous</em> global color if not explicitly set.",
-            .Type         = SettingSpec::SettingType::Color,
-            .Section      = kSectionAppearance,
-            .ApiPath      = "device.secondColor"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = ApplyGlobalColorsTag,
-            .FriendlyName = "(Re)apply global color",
-            .Description  = "You can use this to \"reselect\" and apply the current global color, to force the composition of the derived "
-                            "global palette. This checkbox is ignored if the \"Clear global color\" checkbox is selected.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Access       = SettingSpec::SettingAccess::WriteOnly,
-            .Section      = kSectionAppearance,
-            .ApiPath      = "device.applyGlobalColors"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = ClearGlobalColorTag,
-            .FriendlyName = "Clear global color",
-            .Description  = "Stop applying the global color/derived palette. This takes precedence over the \"(Re)apply global color\" checkbox.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Access       = SettingSpec::SettingAccess::WriteOnly,
-            .Section      = kSectionAppearance,
-            .ApiPath      = "device.clearGlobalColor"
-        }));
-
-        #if SHOW_VU_METER
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = ShowVUMeterTag,
-            .FriendlyName = "Show VU meter",
-            .Description  = "Used to show (checked) or hide the VU meter at the top of the matrix.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionAppearance
-        }));
-        #endif
-
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = RememberCurrentEffectTag,
-            .FriendlyName = "Remember current effect",
-            .Description  = "Indicates if the current effect index should be saved after an effect transition, so the device resumes "
-                            "from the same effect when restarted. Enabling this will lead to more wear on the flash chip of your device.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionAppearance,
-            .ApiPath      = "device.rememberCurrentEffect"
-        }));
-
-        // ---- topology section ----------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = MatrixWidthTag,
-            .FriendlyName = "Matrix width",
-            .Description  = "Active matrix width. WS281x builds validate this by total LED capacity, so width * height must stay within the compiled LED budget.",
-            .Type         = SettingSpec::SettingType::PositiveBigInteger,
-            .MinimumValue = 1.0,
-            .MaximumValue = (double)GetCompiledLEDCount(),
-            .Section      = kSectionTopology,
-            .Priority     = 0,
-            .ApiPath      = "topology.width"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = MatrixHeightTag,
-            .FriendlyName = "Matrix height",
-            .Description  = "Active matrix height. WS281x builds validate this by total LED capacity, so width * height must stay within the compiled LED budget.",
-            .Type         = SettingSpec::SettingType::PositiveBigInteger,
-            .MinimumValue = 1.0,
-            .MaximumValue = (double)GetCompiledLEDCount(),
-            .Section      = kSectionTopology,
-            .Priority     = 1,
-            .ApiPath      = "topology.height"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name         = MatrixSerpentineTag,
-            .FriendlyName = "Serpentine layout",
-            .Description  = "Controls the logical XY mapping for strip-based matrices. HUB75 ignores this because its panel mapping is build-defined.",
-            .Type         = SettingSpec::SettingType::Boolean,
-            .Section      = kSectionTopology,
-            .Priority     = 2,
-            .ApiPath      = "topology.serpentine"
-        }));
-
-        // ---- output section -------------------------------------------------
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name              = OutputDriverTag,
-            .FriendlyName      = "Output driver",
-            .Description       = "Runtime-selected driver. If this differs from the firmware's compiled driver, the API reports recompile required.",
-            .Type              = SettingSpec::SettingType::String,
-            .Section           = kSectionOutput,
-            .Priority          = 10,
-            .ApiPath           = "outputs.driver",
-            .Widget            = SettingSpec::WidgetKind::Select,
-            // Friendly labels for the raw driver identifiers the schema exposes.
-            .Options           = SettingSpec::OptionsSource::SchemaPath,
-            .OptionValues      = {"ws281x", "hub75"},
-            .OptionLabels      = {"WS281x", "HUB75"},
-            .OptionsSchemaPath = "outputs.allowedDrivers"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name              = WS281xChannelCountTag,
-            .FriendlyName      = "WS281x channel count",
-            .Description       = "Number of active strip channels within the compiled maximum. Live updates are limited to WS281x builds.",
-            .Type              = SettingSpec::SettingType::PositiveBigInteger,
-            .MinimumValue      = 1.0,
-            .MaximumValue      = (double)GetCompiledChannelCount(),
-            .Section           = kSectionOutput,
-            .Priority          = 11,
-            .ApiPath           = "outputs.ws281x.channelCount",
-            .Widget            = SettingSpec::WidgetKind::Select,
-            // Concrete list at outputs.ws281x.allowedChannelCounts in the
-            // schema, so the UI doesn't have to derive a range itself.
-            .Options           = SettingSpec::OptionsSource::SchemaPath,
-            .OptionsSchemaPath = "outputs.ws281x.allowedChannelCounts"
-        }));
-        settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name              = WS281xColorOrderTag,
-            .FriendlyName      = "WS281x color order",
-            .Description       = "Byte order used when streaming RGB values to WS281x LEDs. This applies live on strip builds and is ignored on HUB75 builds.",
-            .Type              = SettingSpec::SettingType::String,
-            .Section           = kSectionOutput,
-            .Priority          = 12,
-            .ApiPath           = "outputs.ws281x.colorOrder",
-            .Widget            = SettingSpec::WidgetKind::Select,
-            .Options           = SettingSpec::OptionsSource::SchemaPath,
-            .OptionsSchemaPath = "outputs.ws281x.allowedColorOrders"
-        }));
-
-        // ---- per-channel pin specs ------------------------------------------
-        // ws281xPin{N} specs are synthesized here because all the relevant
-        // details (channel count, naming convention, apiPath format, priority
-        // range) are DeviceConfig internals.
-        // pinSpecStrings holds the String backing; settingSpecs owns the specs.
-        // Both are sized up front so no reallocation can invalidate pointers.
-        constexpr size_t kPinStringsPerChannel = 4; // name, friendly, description, apiPath
-        const auto compiledChannelCount = GetCompiledChannelCount();
-        settingSpecs.reserve(settingSpecs.size() + compiledChannelCount);
-        pinSpecStrings.reserve(compiledChannelCount * kPinStringsPerChannel);
-        const auto stableCStr = [](const String& s) { return s.c_str(); };
-
-        for (size_t i = 0; i < compiledChannelCount; ++i)
-        {
-            const auto& nameStr        = pinSpecStrings.emplace_back(str_sprintf("ws281xPin%zu", i));
-            const auto& friendlyStr    = pinSpecStrings.emplace_back(str_sprintf("WS281x pin %zu", i + 1));
-            const auto& descriptionStr = pinSpecStrings.emplace_back(str_sprintf("GPIO assigned to WS281x channel %zu.", i + 1));
-            const auto& apiPathStr     = pinSpecStrings.emplace_back(str_sprintf("outputs.ws281x.pins[%zu]", i));
-
-            settingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-                .Name         = stableCStr(nameStr),
-                .FriendlyName = stableCStr(friendlyStr),
-                .Description  = stableCStr(descriptionStr),
-                .Type         = SettingSpec::SettingType::Integer,
-                .MinimumValue = -1.0,
-                .MaximumValue = 48.0,
-                .Section      = kSectionOutput,
-                .Priority     = 3 + static_cast<int>(i),
-                .ApiPath      = stableCStr(apiPathStr)
-            }));
-        }
-
-        settingSpecReferences.insert(settingSpecReferences.end(), settingSpecs.begin(), settingSpecs.end());
-    }
-
-    return settingSpecReferences;
 }
 
 const String& DeviceConfig::GetTimeZone() const
@@ -894,7 +446,7 @@ DeviceConfig::ValidateResponse DeviceConfig::ValidateAudioInputPin(int pin) cons
     // The settings API now separates "compiled default" from "active value". External I2S mics can
     // move their DIN pin at boot, but the M5 onboard mic path and the current ADC path are still fixed.
     if (!SupportsConfigurableAudioInputPin())
-        return { false, kRecompileNeededMessage };
+        return { false, DeviceConfigInternal::RecompileNeededMessage() };
 
     if (pin == -1)
         return { true, "" };
@@ -1002,517 +554,3 @@ DeviceConfig::ValidateResponse DeviceConfig::ValidateOpenWeatherAPIKey(const Str
 }
 #endif  // ENABLE_WIFI
 
-void DeviceConfig::SetColorSettings(const CRGB& newGlobalColor, const CRGB& newSecondColor)
-{
-    globalColor = newGlobalColor;
-    secondColor = newSecondColor;
-    applyGlobalColors = true;
-
-    SaveToJSON();
-}
-
-// This function contains the logic for dealing with the various color-related settings we have.
-// The logic effectively mimics the behavior of pressing a color button on the IR remote control when (only) the
-// global color is set or (re)applied, but also allows the secondary global palette color to be specified directly.
-// The code in this function figures out how to prioritize and combine the values of (optional) settings; the actual
-// logic for applying the correct color(s) and palette is contained in a number of EffectManager member functions.
-void DeviceConfig::ApplyColorSettings(std::optional<CRGB> newGlobalColor, std::optional<CRGB> newSecondColor, bool clearGlobalColor, bool forceApplyGlobalColor)
-{
-    // If we're asked to clear the global color, we'll remember any colors we were passed, but won't do anything with them
-    if (clearGlobalColor)
-    {
-        if (newGlobalColor.has_value())
-            globalColor = newGlobalColor.value();
-        if (newSecondColor.has_value())
-            secondColor = newSecondColor.value();
-
-        g_ptrSystem->GetEffectManager().ClearRemoteColor();
-
-        applyGlobalColors = false;
-
-        SaveToJSON();
-
-        return;
-    }
-
-    CRGB finalGlobalColor = newGlobalColor.has_value() ? newGlobalColor.value() : globalColor;
-    forceApplyGlobalColor = forceApplyGlobalColor || newGlobalColor.has_value();
-
-    // If we were given a second color, set it and the global one if necessary. Then have EffectManager do its thing...
-    if (newSecondColor.has_value())
-    {
-        if (forceApplyGlobalColor)
-        {
-            applyGlobalColors = true;
-            globalColor = finalGlobalColor;
-        }
-
-        secondColor = newSecondColor.value();
-
-        g_ptrSystem->GetEffectManager().ApplyGlobalPaletteColors();
-
-        SaveToJSON();
-    }
-    else if (forceApplyGlobalColor)
-    {
-        // ...otherwise, apply the "set global color" logic if we were asked to do so
-        g_ptrSystem->GetEffectManager().ApplyGlobalColor(finalGlobalColor);
-    }
-}
-
-DeviceConfig::ValidateResponse DeviceConfig::ValidateTopology(uint16_t width, uint16_t height, bool serpentine) const
-{
-    if (width == 0 || height == 0)
-        return { false, "matrix dimensions must be greater than zero" };
-
-    // The strip path sizes its live buffers from total LED capacity, not the original compile-time aspect ratio.
-    // That keeps reshaping flexible while still refusing requests that would outgrow the compiled backing store.
-    if (static_cast<size_t>(width) * height > GetCompiledLEDCount())
-        return { false, kRecompileNeededMessage };
-
-    if (IsHub75Build())
-    {
-        if (width != GetCompiledMatrixWidth() || height != GetCompiledMatrixHeight())
-            return { false, kRecompileNeededMessage };
-
-        if (serpentine != GetCompiledMatrixSerpentine())
-            return { false, kRecompileNeededMessage };
-    }
-
-    return { true, "" };
-}
-
-DeviceConfig::ValidateResponse DeviceConfig::ValidateOutputDriver(OutputDriver driver) const
-{
-    if (driver != GetCompiledOutputDriver())
-        return { false, kRecompileNeededMessage };
-
-    return { true, "" };
-}
-
-DeviceConfig::ValidateResponse DeviceConfig::ValidateWS281xSettings(size_t channelCount, const std::array<int8_t, NUM_CHANNELS>& pins, WS281xColorOrder colorOrder) const
-{
-    if (channelCount == 0)
-        return { false, "channel count must be greater than zero" };
-
-    if (channelCount > GetCompiledChannelCount())
-        return { false, kRecompileNeededMessage };
-
-    if (IsHub75Build())
-    {
-        if (channelCount != GetCompiledChannelCount())
-            return { false, kRecompileNeededMessage };
-
-        if (pins != GetCompiledWS281xPins())
-            return { false, kRecompileNeededMessage };
-
-        if (colorOrder != GetCompiledWS281xColorOrder())
-            return { false, kRecompileNeededMessage };
-    }
-
-    for (size_t i = 0; i < channelCount; ++i)
-    {
-        if (pins[i] < 0)
-            return { false, "active channels require valid GPIO pins" };
-
-        if (!GPIO_IS_VALID_OUTPUT_GPIO(static_cast<gpio_num_t>(pins[i])))
-            return { false, "WS281x channel pins must be valid output GPIOs" };
-
-        for (size_t j = i + 1; j < channelCount; ++j)
-        {
-            if (pins[i] == pins[j])
-                return { false, "WS281x channel pins must be unique" };
-        }
-    }
-
-    return { true, "" };
-}
-
-DeviceConfig::ValidateResponse DeviceConfig::ValidateRuntimeConfig(const RuntimeConfig& config) const
-{
-    auto [driverValid, driverMessage] = ValidateOutputDriver(config.outputs.driver);
-    if (!driverValid)
-        return { false, driverMessage };
-
-    auto [topologyValid, topologyMessage] = ValidateTopology(config.topology.width, config.topology.height, config.topology.serpentine);
-    if (!topologyValid)
-        return { false, topologyMessage };
-
-    auto [ws281xValid, ws281xMessage] = ValidateWS281xSettings(config.outputs.channelCount, config.outputs.outputPins, config.outputs.colorOrder);
-    if (!ws281xValid)
-        return { false, ws281xMessage };
-
-    return { true, "" };
-}
-
-bool DeviceConfig::SetRuntimeConfig(const RuntimeConfig& config, bool skipWrite, String* errorMessage)
-{
-    auto [isValid, validationMessage] = ValidateRuntimeConfig(config);
-    if (!isValid)
-    {
-        if (errorMessage)
-            *errorMessage = validationMessage;
-        return false;
-    }
-
-    const bool changed =
-        runtimeTopology.width != config.topology.width
-        || runtimeTopology.height != config.topology.height
-        || runtimeTopology.serpentine != config.topology.serpentine
-        || runtimeOutputs.driver != config.outputs.driver
-        || runtimeOutputs.channelCount != config.outputs.channelCount
-        || runtimeOutputs.outputPins != config.outputs.outputPins
-        || runtimeOutputs.colorOrder != config.outputs.colorOrder;
-
-    runtimeTopology = config.topology;
-    runtimeOutputs = config.outputs;
-
-    if (!skipWrite)
-        SaveToJSON();
-
-    if (changed && !skipWrite)
-        LogRuntimeConfig("runtime config changed");
-
-    if (errorMessage)
-        *errorMessage = "";
-
-    return true;
-}
-
-void DeviceConfig::SerializeUnifiedSettings(JsonObject root) const
-{
-    auto device = root["device"].to<JsonObject>();
-    device["hostname"] = GetHostname();
-    device["location"] = GetLocation();
-    device["locationIsZip"] = IsLocationZip();
-    device["countryCode"] = GetCountryCode();
-    device["timeZone"] = GetTimeZone();
-    device["use24HourClock"] = Use24HourClock();
-    device["useCelsius"] = UseCelsius();
-    device["ntpServer"] = GetNTPServer();
-    device["rememberCurrentEffect"] = RememberCurrentEffect();
-    device["powerLimit"] = GetPowerLimit();
-    device["brightness"] = GetBrightness();
-    device["globalColor"] = GlobalColor();
-    device["secondColor"] = SecondColor();
-    device["applyGlobalColors"] = ApplyGlobalColors();
-
-    #if ENABLE_REMOTE
-    auto remote = device["remote"].to<JsonObject>();
-    remote["enabled"] = true;
-    remote["pin"] = IR_REMOTE_PIN;
-    #else
-    auto remote = device["remote"].to<JsonObject>();
-    remote["enabled"] = false;
-    remote["pin"] = -1;
-    #endif
-
-    auto audio = device["audio"].to<JsonObject>();
-    audio["enabled"] =
-    #if ENABLE_AUDIO
-        true;
-    #else
-        false;
-    #endif
-    audio["audioInputPin"] = GetAudioInputPin();
-    audio["compiledDefaultPin"] = GetCompiledAudioInputPin();
-    audio["mode"] = GetAudioInputModeName();
-    audio["liveApply"] = SupportsLiveAudioInputReconfigure();
-    audio["requiresReboot"] = !SupportsLiveAudioInputReconfigure();
-    audio["supportsPinOverride"] = SupportsConfigurableAudioInputPin();
-
-    auto topology = root["topology"].to<JsonObject>();
-    topology["width"] = GetMatrixWidth();
-    topology["height"] = GetMatrixHeight();
-    topology["serpentine"] = IsMatrixSerpentine();
-    topology["ledCount"] = GetActiveLEDCount();
-    topology["liveApply"] = SupportsLiveTopology();
-
-    auto outputs = root["outputs"].to<JsonObject>();
-    outputs["driver"] = GetRuntimeDriverName();
-    outputs["compiledDriver"] = GetCompiledDriverName();
-    outputs["liveApply"] = SupportsLiveOutputReconfigure();
-
-    auto ws281x = outputs["ws281x"].to<JsonObject>();
-    ws281x["channelCount"] = GetChannelCount();
-    ws281x["compiledMaxChannels"] = GetCompiledChannelCount();
-    ws281x["colorOrder"] = GetColorOrderName(GetWS281xColorOrder());
-    ws281x["compiledColorOrder"] = GetColorOrderName(GetCompiledWS281xColorOrder());
-    AppendPins(ws281x["pins"].to<JsonArray>(), GetWS281xPins());
-}
-
-void DeviceConfig::SerializeUnifiedSettingsSchema(JsonObject root) const
-{
-    auto topology = root["topology"].to<JsonObject>();
-    topology["compiledMaxWidth"] =
-        GetCompiledOutputDriver() == OutputDriver::HUB75
-            ? GetCompiledMatrixWidth()
-            : GetCompiledLEDCount();
-    topology["compiledMaxHeight"] =
-        GetCompiledOutputDriver() == OutputDriver::HUB75
-            ? GetCompiledMatrixHeight()
-            : GetCompiledLEDCount();
-    topology["compiledNominalWidth"] = GetCompiledMatrixWidth();
-    topology["compiledNominalHeight"] = GetCompiledMatrixHeight();
-    topology["compiledMaxLEDs"] = GetCompiledLEDCount();
-    topology["liveApply"] = SupportsLiveTopology();
-    topology["rejectMessage"] = "recompile needed";
-
-    auto outputs = root["outputs"].to<JsonObject>();
-    outputs["compiledDriver"] = GetCompiledDriverName();
-    outputs["liveApply"] = SupportsLiveOutputReconfigure();
-    outputs["rejectMessage"] = "recompile needed";
-
-    auto drivers = outputs["allowedDrivers"].to<JsonArray>();
-    drivers.add(GetCompiledDriverName());
-
-    auto ws281x = outputs["ws281x"].to<JsonObject>();
-    const auto compiledChannels = GetCompiledChannelCount();
-    ws281x["compiledMaxChannels"] = compiledChannels;
-    ws281x["compiledMaxLEDs"] = GetCompiledLEDCount();
-    ws281x["compiledColorOrder"] = GetColorOrderName(GetCompiledWS281xColorOrder());
-    auto allowedChannelCounts = ws281x["allowedChannelCounts"].to<JsonArray>();
-    for (size_t channel = 1; channel <= compiledChannels; ++channel)
-        allowedChannelCounts.add(channel);
-
-    auto allowedColorOrders = ws281x["allowedColorOrders"].to<JsonArray>();
-    allowedColorOrders.add("RGB");
-    allowedColorOrders.add("RBG");
-    allowedColorOrders.add("GRB");
-    allowedColorOrders.add("GBR");
-    allowedColorOrders.add("BRG");
-    allowedColorOrders.add("BGR");
-    AppendPins(ws281x["compiledPins"].to<JsonArray>(), GetCompiledWS281xPins());
-
-    auto device = root["device"].to<JsonObject>();
-    auto remote = device["remote"].to<JsonObject>();
-    remote["enabled"] =
-    #if ENABLE_REMOTE
-        true;
-    #else
-        false;
-    #endif
-    remote["pin"] = IR_REMOTE_PIN;
-
-    auto audio = device["audio"].to<JsonObject>();
-    audio["enabled"] =
-    #if ENABLE_AUDIO
-        true;
-    #else
-        false;
-    #endif
-    audio["compiledDefaultPin"] = GetCompiledAudioInputPin();
-    audio["mode"] = GetAudioInputModeName();
-    audio["liveApply"] = SupportsLiveAudioInputReconfigure();
-    audio["requiresReboot"] = !SupportsLiveAudioInputReconfigure();
-    audio["supportsPinOverride"] = SupportsConfigurableAudioInputPin();
-    audio["rejectMessage"] = "recompile needed";
-
-    struct SectionInfo
-    {
-        const char* id;
-        const char* title;
-        const char* description;
-    };
-
-    static constexpr SectionInfo kSections[] =
-    {
-        { "topology",   "Topology",          "Active matrix dimensions and layout." },
-        { "output",     "Output",            "LED driver, channel count, color order, and per-channel pin assignments." },
-        { "appearance", "Appearance",        "Brightness, colors, effect rotation, and visual preferences." },
-        { "audio",      "Audio",             "Microphone input pin and audio capture configuration." },
-        { "location",   "Location",          "Where the device is for weather and timezone defaults." },
-        { "clock",      "Clock & Weather",   "Time display, NTP, and weather API options." },
-        { "system",     "System",            "Identification, power limits, and other system-wide options." },
-    };
-
-    auto sections = root["sections"].to<JsonArray>();
-    for (const auto& info : kSections)
-    {
-        auto entry = sections.add<JsonObject>();
-        entry["id"] = info.id;
-        entry["title"] = info.title;
-        entry["description"] = info.description;
-    }
-}
-
-DeviceConfig::ValidateResponse DeviceConfig::ParseAndValidateUnifiedSettings(JsonObjectConst root, UnifiedSettingsRequest& out) const
-{
-    out = UnifiedSettingsRequest{};
-    out.requestedRuntimeConfig = GetRuntimeConfig();
-
-    if (root["topology"].is<JsonObjectConst>())
-    {
-        auto topology = root["topology"].as<JsonObjectConst>();
-        if (topology["width"].is<uint16_t>())
-        {
-            out.requestedRuntimeConfig.topology.width = topology["width"].as<uint16_t>();
-            out.runtimeConfigTouched = true;
-        }
-        if (topology["height"].is<uint16_t>())
-        {
-            out.requestedRuntimeConfig.topology.height = topology["height"].as<uint16_t>();
-            out.runtimeConfigTouched = true;
-        }
-        if (topology["serpentine"].is<bool>())
-        {
-            out.requestedRuntimeConfig.topology.serpentine = topology["serpentine"].as<bool>();
-            out.runtimeConfigTouched = true;
-        }
-    }
-
-    if (root["outputs"].is<JsonObjectConst>())
-    {
-        auto outputs = root["outputs"].as<JsonObjectConst>();
-        if (outputs["driver"].is<String>())
-        {
-            const auto driver = outputs["driver"].as<String>();
-            out.requestedRuntimeConfig.outputs.driver = driver == "hub75" ? OutputDriver::HUB75 : OutputDriver::WS281x;
-            out.runtimeConfigTouched = true;
-        }
-
-        if (outputs["ws281x"].is<JsonObjectConst>())
-        {
-            auto ws281x = outputs["ws281x"].as<JsonObjectConst>();
-            if (ws281x["channelCount"].is<size_t>())
-            {
-                out.requestedRuntimeConfig.outputs.channelCount = ws281x["channelCount"].as<size_t>();
-                out.runtimeConfigTouched = true;
-            }
-            if (ws281x["colorOrder"].is<String>())
-            {
-                if (auto colorOrder = ParseWS281xColorOrderName(ws281x["colorOrder"].as<String>()); colorOrder.has_value())
-                {
-                    out.requestedRuntimeConfig.outputs.colorOrder = colorOrder.value();
-                    out.runtimeConfigTouched = true;
-                }
-            }
-            if (ws281x["pins"].is<JsonArrayConst>())
-            {
-                auto pins = ws281x["pins"].as<JsonArrayConst>();
-                for (size_t i = 0; i < out.requestedRuntimeConfig.outputs.outputPins.size() && i < pins.size(); ++i)
-                {
-                    if (pins[i].is<int>())
-                    {
-                        out.requestedRuntimeConfig.outputs.outputPins[i] = pins[i].as<int>();
-                        out.runtimeConfigTouched = true;
-                    }
-                }
-            }
-        }
-    }
-
-    auto [runtimeConfigValid, runtimeConfigMessage] = ValidateRuntimeConfig(out.requestedRuntimeConfig);
-    if (!runtimeConfigValid)
-        return { false, runtimeConfigMessage };
-
-    if (root["device"].is<JsonObjectConst>())
-    {
-        auto device = root["device"].as<JsonObjectConst>();
-
-        FieldAccess::AssignIfPresent(device, HostnameTag, out.hostname);
-        FieldAccess::AssignIfPresent(device, LocationTag, out.location);
-        FieldAccess::AssignIfPresent(device, LocationIsZipTag, out.locationIsZip);
-        FieldAccess::AssignIfPresent(device, CountryCodeTag, out.countryCode);
-        FieldAccess::AssignIfPresent(device, TimeZoneTag, out.timeZone);
-        FieldAccess::AssignIfPresent(device, Use24HourClockTag, out.use24HourClock);
-        FieldAccess::AssignIfPresent(device, UseCelsiusTag, out.useCelsius);
-        FieldAccess::AssignIfPresent(device, NTPServerTag, out.ntpServer);
-        FieldAccess::AssignIfPresent(device, RememberCurrentEffectTag, out.rememberCurrentEffect);
-
-        if (device[OpenWeatherApiKeyTag].is<String>())
-        {
-            const auto requestedKey = device[OpenWeatherApiKeyTag].as<String>();
-            auto [isValid, validationMessage] = ValidateOpenWeatherAPIKey(requestedKey);
-            if (!isValid)
-                return { false, validationMessage };
-
-            out.openWeatherApiKey = requestedKey;
-        }
-
-        if (device[PowerLimitTag].is<int>())
-        {
-            const int requestedPowerLimit = device[PowerLimitTag].as<int>();
-            auto [isValid, validationMessage] = ValidatePowerLimit(requestedPowerLimit);
-            if (!isValid)
-                return { false, validationMessage };
-
-            out.powerLimit = requestedPowerLimit;
-        }
-
-        if (device[BrightnessTag].is<int>())
-        {
-            const int requestedBrightness = device[BrightnessTag].as<int>();
-            auto [isValid, validationMessage] = ValidateBrightness(requestedBrightness);
-            if (!isValid)
-                return { false, validationMessage };
-
-            out.brightness = requestedBrightness;
-        }
-
-        String audioErrorMessage;
-        const auto requestedAudioInputPin = ResolveUnifiedAudioInputPin(device, &audioErrorMessage);
-        if (!audioErrorMessage.isEmpty())
-            return { false, audioErrorMessage };
-        if (requestedAudioInputPin.has_value())
-        {
-            auto [isValid, validationMessage] = ValidateAudioInputPin(requestedAudioInputPin.value());
-            if (!isValid)
-                return { false, validationMessage };
-
-            out.audioInputPin = requestedAudioInputPin.value();
-        }
-
-        FieldAccess::AssignIfPresent(device, GlobalColorTag, out.globalColor);
-        FieldAccess::AssignIfPresent(device, SecondColorTag, out.secondColor);
-        out.clearGlobalColor = device[ClearGlobalColorTag].is<bool>() && device[ClearGlobalColorTag].as<bool>();
-        out.applyGlobalColors = device[ApplyGlobalColorsTag].is<bool>() && device[ApplyGlobalColorsTag].as<bool>();
-    }
-
-    return { true, "" };
-}
-
-bool DeviceConfig::ApplyUnifiedDeviceSettings(const UnifiedSettingsRequest& request, String* errorMessage)
-{
-    FieldAccess::ApplyIfPresent(request.hostname, *this, &DeviceConfig::SetHostname);
-    FieldAccess::ApplyIfPresent(request.location, *this, &DeviceConfig::SetLocation);
-    FieldAccess::ApplyIfPresent(request.locationIsZip, *this, &DeviceConfig::SetLocationIsZip);
-    FieldAccess::ApplyIfPresent(request.countryCode, *this, &DeviceConfig::SetCountryCode);
-    FieldAccess::ApplyIfPresent(request.openWeatherApiKey, *this, &DeviceConfig::SetOpenWeatherAPIKey);
-    FieldAccess::ApplyIfPresent(request.timeZone, [this](const String& value) { SetTimeZone(value); });
-    FieldAccess::ApplyIfPresent(request.use24HourClock, *this, &DeviceConfig::Set24HourClock);
-    FieldAccess::ApplyIfPresent(request.useCelsius, *this, &DeviceConfig::SetUseCelsius);
-    FieldAccess::ApplyIfPresent(request.ntpServer, *this, &DeviceConfig::SetNTPServer);
-    FieldAccess::ApplyIfPresent(request.rememberCurrentEffect, *this, &DeviceConfig::SetRememberCurrentEffect);
-    FieldAccess::ApplyIfPresent(request.powerLimit, *this, &DeviceConfig::SetPowerLimit);
-    FieldAccess::ApplyIfPresent(request.brightness, *this, &DeviceConfig::SetBrightness);
-
-    if (request.audioInputPin.has_value())
-    {
-        const int oldPin = GetAudioInputPin();
-        SetAudioInputPin(request.audioInputPin.value());
-
-        const int newPin = GetAudioInputPin();
-        if (newPin != oldPin && SupportsLiveAudioInputReconfigure() && g_ptrSystem && g_ptrSystem->HasAudioService())
-        {
-            auto& audioService = g_ptrSystem->GetAudioService();
-            if (!audioService.Reconfigure(AudioConfig::FromCurrentSettings()))
-            {
-                SetAudioInputPin(oldPin);
-                if (errorMessage)
-                    *errorMessage = "Audio input pin live apply failed";
-                return false;
-            }
-        }
-    }
-
-    ApplyColorSettings(request.globalColor,
-                       request.secondColor,
-                       request.clearGlobalColor,
-                       request.applyGlobalColors);
-
-    if (errorMessage)
-        *errorMessage = "";
-
-    return true;
-}
