+++ src/systemcontainer.cpp
@@ -302,7 +302,7 @@ int SystemContainer::GetConfiguredAudioInputPin() const
     return DeviceConfig::GetCompiledAudioInputPin();
 }
 
-bool SystemContainer::ApplyRuntimeConfiguration(String* errorMessage)
+DeviceConfig::ValidateResponse SystemContainer::ApplyRuntimeConfiguration()
 {
     // Runtime topology changes retarget both the graphics objects and the
     // WiFi frame buffers. Lock all three domains together so the render task
@@ -313,11 +313,7 @@ bool SystemContainer::ApplyRuntimeConfiguration(String* errorMessage)
     auto& config = GetDeviceConfig();
 
     if (config.RequiresRecompileForCurrentRuntimeConfig())
-    {
-        if (errorMessage)
-            *errorMessage = "recompile needed";
-        return false;
-    }
+        return { false, "recompile needed" };
 
     #if USE_WS281X
     try
@@ -343,54 +339,52 @@ bool SystemContainer::ApplyRuntimeConfiguration(String* errorMessage)
 
         // The mutable WS281x output layer is now always owned by NightDriver so live pin/count/color-order
         // changes stay on one transport path instead of handing off between different ESP32 RMT backends.
-        if (_ptrDevices && !SetupWS281xOutputManager().ApplyConfig(config, *_ptrDevices, errorMessage))
-            return false;
+        if (_ptrDevices)
+        {
+            String applyConfigError;
+            if (!SetupWS281xOutputManager().ApplyConfig(config, *_ptrDevices, &applyConfigError))
+                return { false, applyConfigError };
+        }
 
         if (_ptrEffectManager && !_ptrEffectManager->ReinitializeEffects())
-        {
-            if (errorMessage)
-                *errorMessage = "failed to reinitialize effects for new topology";
-            return false;
-        }
+            return { false, "failed to reinitialize effects for new topology" };
     }
     catch (const std::bad_alloc&)
     {
         debugE("Runtime configuration failed due to insufficient memory");
-        if (errorMessage)
-            *errorMessage = "insufficient memory for runtime reconfiguration";
-        return false;
+        return { false, "insufficient memory for runtime reconfiguration" };
     }
     #endif
 
-    if (errorMessage)
-        *errorMessage = "";
-
-    return true;
+    return { true, "" };
 }
 
-bool SystemContainer::ApplyRuntimeConfigurationTransaction(const DeviceConfig::RuntimeConfig& requestedConfig, String* errorMessage)
+DeviceConfig::ValidateResponse SystemContainer::ApplyRuntimeConfigurationTransaction(const DeviceConfig::RuntimeConfig& requestedConfig)
 {
     auto& deviceConfig = GetDeviceConfig();
     const auto previousConfig = deviceConfig.GetRuntimeConfig();
 
-    if (!deviceConfig.SetRuntimeConfig(requestedConfig, true, errorMessage))
-        return false;
+    auto [stagedConfigValid, stagedConfigMessage] = deviceConfig.SetRuntimeConfig(requestedConfig, true);
+    if (!stagedConfigValid)
+        return { false, stagedConfigMessage };
 
-    if (!ApplyRuntimeConfiguration(errorMessage))
+    auto [runtimeConfigApplied, runtimeErrorMessage] = ApplyRuntimeConfiguration();
+    if (!runtimeConfigApplied)
     {
-        deviceConfig.SetRuntimeConfig(previousConfig, true, nullptr);
+        (void)deviceConfig.SetRuntimeConfig(previousConfig, true);
 
-        String rollbackError;
-        if (!ApplyRuntimeConfiguration(&rollbackError))
+        auto [rollbackApplied, rollbackError] = ApplyRuntimeConfiguration();
+        if (!rollbackApplied)
             debugE("Failed to roll back runtime configuration after apply error: %s", rollbackError.c_str());
 
-        return false;
+        return { false, runtimeErrorMessage };
     }
 
-    if (!deviceConfig.SetRuntimeConfig(requestedConfig, false, errorMessage))
-        return false;
+    auto [persistedConfigValid, persistedConfigMessage] = deviceConfig.SetRuntimeConfig(requestedConfig, false);
+    if (!persistedConfigValid)
+        return { false, persistedConfigMessage };
 
-    return true;
+    return { true, "" };
 }
 
 #if ENABLE_WIFI
