+++ src/webserver.cpp
@@ -38,26 +38,19 @@
 #include <FS.h>
 #include <algorithm>
 #include <cstdlib>
-#include <memory>
-#include <mutex>
+#include <limits>
 #include <utility>
 
-#include "audioservice.h"
 #include "deviceconfig.h"
+#include "effects.h"
 #include "effectmanager.h"
 #include "gfxbase.h"
 #include "improvserial.h"
-#include "ledstripeffect.h"
 #include "soundanalyzer.h"
 #include "systemcontainer.h"
 #include "taskmgr.h"
 #include "values.h"
 #include "websocketserver.h"   // Stop() tears down our websocket clients first
-
-namespace
-{
-    std::recursive_mutex g_settingSpecsCacheMutex;
-}
 
 // Static member initializers
 
@@ -166,40 +159,6 @@ void CWebServer::Stop()
     _server.end();
     _running.store(false);
     debugI("WebServer: stop completed");
-}
-
-// ApplyAudioInputPinChange - factored out of SetSettingsIfPresent /
-// SetUnifiedSettings. Compares the current persisted pin against `oldPin`;
-// if the pin actually changed and the build supports a live reconfigure
-// (and AudioService is present), pushes the new AudioConfig through
-// AudioService::Reconfigure(). Reverts the persisted pin on failure so the
-// caller's serialized response matches the live state. Safe to call when
-// g_ptrSystem is null or AudioService isn't constructed.
-bool CWebServer::ApplyAudioInputPinChange(int oldPin)
-{
-    if (!g_ptrSystem || !g_ptrSystem->HasDeviceConfig())
-        return true;
-
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    const int newPin = deviceConfig.GetAudioInputPin();
-
-    if (newPin == oldPin)
-        return true;
-
-    if (!deviceConfig.SupportsLiveAudioInputReconfigure())
-        return true;
-
-    if (!g_ptrSystem->HasAudioService())
-        return true;
-
-    auto& audioService = g_ptrSystem->GetAudioService();
-    if (!audioService.Reconfigure(AudioConfig::FromCurrentSettings()))
-    {
-        debugW("Audio: live reconfigure failed; reverting persisted pin to %d", oldPin);
-        deviceConfig.SetAudioInputPin(oldPin);
-        return false;
-    }
-    return true;
 }
 
 // begin - register page load handlers and start serving pages
@@ -356,50 +315,6 @@ void CWebServer::SendBufferOverflowResponse(AsyncWebServerRequest * pRequest)
     );
 }
 
-void CWebServer::GetEffectListText(AsyncWebServerRequest * pRequest)
-{
-    debugV("GetEffectListText");
-
-    auto response = std::make_unique<AsyncJsonResponse>();
-    auto& j = response->getRoot();
-    auto& effectManager = g_ptrSystem->GetEffectManager();
-    bool overflow = false;
-
-    {
-        std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-
-        j["currentEffect"]         = effectManager.GetCurrentEffectIndex();
-        j["millisecondsRemaining"] = effectManager.GetTimeRemainingForCurrentEffect();
-        j["eternalInterval"]       = effectManager.IsIntervalEternal();
-        j["effectInterval"]        = effectManager.GetInterval();
-
-        for (const auto& effect : effectManager.EffectsList())
-        {
-            auto effectDoc = CreateJsonDocument();
-
-            effectDoc["name"]    = effect->FriendlyName();
-            effectDoc["enabled"] = effect->IsEnabled();
-            effectDoc["core"]    = effect->IsCoreEffect();
-
-            if (!j["Effects"].add(effectDoc))
-            {
-                debugV("JSON response buffer overflow!");
-                overflow = true;
-                break;
-            }
-        }
-    }
-
-    if (overflow)
-    {
-        response.reset();
-        SendBufferOverflowResponse(pRequest);
-        return;
-    }
-
-    AddCORSHeaderAndSendResponse(pRequest, response.release());
-}
-
 void CWebServer::GetStatistics(AsyncWebServerRequest * pRequest, StatisticsType statsType) const
 {
     debugV("GetStatistics");
@@ -466,754 +381,6 @@ void CWebServer::GetStatistics(AsyncWebServerRequest * pRequest, StatisticsType 
     AddCORSHeaderAndSendResponse(pRequest, response);
 }
 
-void CWebServer::SetCurrentEffectIndex(AsyncWebServerRequest * pRequest)
-{
-    debugV("SetCurrentEffectIndex");
-    PushPostParamIfPresent<size_t>(pRequest, "currentEffectIndex", SET_VALUE(g_ptrSystem->GetEffectManager().SetCurrentEffectIndex(value)));
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::EnableEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("EnableEffect");
-    PushPostParamIfPresent<size_t>(pRequest, "effectIndex", SET_VALUE(g_ptrSystem->GetEffectManager().EnableEffect(value)));
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::DisableEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("DisableEffect");
-    PushPostParamIfPresent<size_t>(pRequest, "effectIndex", SET_VALUE(g_ptrSystem->GetEffectManager().DisableEffect(value)));
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::MoveEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("MoveEffect");
-
-    auto fromIndex = GetEffectIndexFromParam(pRequest, true);
-    if (fromIndex == -1)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return;
-    }
-
-    PushPostParamIfPresent<size_t>(pRequest, "newIndex", SET_VALUE(g_ptrSystem->GetEffectManager().MoveEffect(fromIndex, value)));
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::CopyEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("CopyEffect");
-
-    auto index = GetEffectIndexFromParam(pRequest, true);
-    if (index == -1)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return;
-    }
-
-    auto effect = g_ptrSystem->GetEffectManager().CopyEffect(index);
-    if (!effect)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return;
-    }
-
-    ApplyEffectSettings(pRequest, effect);
-
-    if (g_ptrSystem->GetEffectManager().AppendEffect(effect))
-        SendEffectSettingsResponse(pRequest, effect);
-    else
-        AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::DeleteEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("DeleteEffect");
-
-    auto index = GetEffectIndexFromParam(pRequest, true);
-    if (index == -1)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return;
-    }
-
-    if (g_ptrSystem->GetEffectManager().IsCoreEffect(index))
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, "Can't delete core effect");
-        return;
-    }
-
-    g_ptrSystem->GetEffectManager().DeleteEffect(index);
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::NextEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("NextEffect");
-    g_ptrSystem->GetEffectManager().NextEffect();
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::PreviousEffect(AsyncWebServerRequest * pRequest)
-{
-    debugV("PreviousEffect");
-    g_ptrSystem->GetEffectManager().PreviousEffect();
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
-
-void CWebServer::SendSettingSpecsResponse(AsyncWebServerRequest * pRequest, const std::vector<std::reference_wrapper<SettingSpec>> & settingSpecs)
-{
-    String json;
-    if (!BuildSettingSpecsJson(json, settingSpecs))
-    {
-        debugW("Unable to build setting specs JSON response.");
-        SendBufferOverflowResponse(pRequest);
-        return;
-    }
-
-    AddCORSHeaderAndSendResponse(pRequest, pRequest->beginResponse(HttpOk, "application/json", json));
-}
-
-bool CWebServer::BuildSettingSpecsJson(String& json, const std::vector<std::reference_wrapper<SettingSpec>> & settingSpecs)
-{
-    auto jsonDoc = CreateJsonDocument();
-    auto jsonArray = jsonDoc.to<JsonArray>();
-
-    for (const auto& specWrapper : settingSpecs)
-    {
-        const auto& spec = specWrapper.get();
-        auto specObject = jsonArray.add<JsonObject>();
-        if (specObject.isNull())
-        {
-            debugW("Setting specs JSON object allocation failed.");
-            return false;
-        }
-
-        specObject["name"] = spec.Name;
-        specObject["friendlyName"] = spec.FriendlyName;
-        if (spec.Description)
-            specObject["description"] = spec.Description;
-        specObject["type"] = to_value(spec.Type);
-        specObject["typeName"] = spec.TypeName();
-        if (spec.HasValidation)
-            specObject["hasValidation"] = true;
-        if (spec.MinimumValue.has_value())
-            specObject["minimumValue"] = spec.MinimumValue.value();
-        if (spec.MaximumValue.has_value())
-            specObject["maximumValue"] = spec.MaximumValue.value();
-        if (spec.EmptyAllowed.has_value())
-            specObject["emptyAllowed"] = spec.EmptyAllowed.value();
-        switch (spec.Access)
-        {
-            case SettingSpec::SettingAccess::ReadOnly:
-                specObject["readOnly"] = true;
-                break;
-
-            case SettingSpec::SettingAccess::WriteOnly:
-                specObject["writeOnly"] = true;
-                break;
-
-            default:
-                // Default is read/write, so we don't need to specify that
-                break;
-        }
-
-        // ---- UI metadata: section, priority, requiresReboot, apiPath, widget ----
-
-        if (spec.Section)
-            specObject["section"] = spec.Section;
-        if (spec.Priority.has_value())
-            specObject["priority"] = spec.Priority.value();
-        if (spec.RequiresReboot)
-            specObject["requiresReboot"] = true;
-        if (spec.ApiPath)
-            specObject["apiPath"] = spec.ApiPath;
-
-        if (spec.Widget != SettingSpec::WidgetKind::Default)
-        {
-            auto widget = specObject["widget"].to<JsonObject>();
-            widget["kind"] = spec.WidgetName();
-
-            if (spec.Widget == SettingSpec::WidgetKind::Slider
-                && spec.DisplayRawMin.has_value()) // Presence of other Display members is validated at SettingSpec construction
-            {
-                auto scale = widget["displayScale"].to<JsonObject>();
-                scale["rawMin"]     = spec.DisplayRawMin.value();
-                scale["rawMax"]     = spec.DisplayRawMax.value();
-                scale["displayMin"] = spec.DisplayMin.value();
-                scale["displayMax"] = spec.DisplayMax.value();
-                if (spec.DisplaySuffix)
-                    scale["suffix"] = spec.DisplaySuffix;
-            }
-            else if (spec.Widget == SettingSpec::WidgetKind::IntervalToggle)
-            {
-                auto interval = widget["interval"].to<JsonObject>();
-                interval["unitDivisor"] = spec.IntervalUnitDivisor;
-                if (spec.IntervalUnitLabel)
-                    interval["unitLabel"] = spec.IntervalUnitLabel;
-                if (spec.IntervalOnLabel)
-                    interval["onLabel"] = spec.IntervalOnLabel;
-                if (spec.IntervalOffLabel)
-                    interval["offLabel"] = spec.IntervalOffLabel;
-            }
-            else if (spec.Widget == SettingSpec::WidgetKind::Select)
-            {
-                auto options = widget["options"].to<JsonObject>();
-                options["source"] = spec.OptionsSourceName();
-
-                auto addOptionArrays = [&]()
-                {
-                    auto valuesArr = options["values"].to<JsonArray>();
-                    for (auto entry : spec.OptionValues)
-                        valuesArr.add(entry ? entry : "");
-                    auto labelsArr = options["labels"].to<JsonArray>();
-                    for (auto entry : spec.OptionLabels)
-                        labelsArr.add(entry ? entry : "");
-                };
-
-                if (spec.Options == SettingSpec::OptionsSource::Inline)
-                {
-                    addOptionArrays();
-                }
-                else if (spec.OptionsSchemaPath)
-                {
-                    options["schemaPath"] = spec.OptionsSchemaPath;
-                    if (!spec.OptionValues.empty())
-                        addOptionArrays();
-                }
-
-                if (spec.OptionsExternalUrl)
-                    options["url"] = spec.OptionsExternalUrl;
-            }
-        }
-    }
-
-    if (jsonDoc.overflowed())
-    {
-        debugW("Setting specs JSON document overflowed.");
-        return false;
-    }
-
-    const size_t measuredLength = measureJson(jsonArray);
-    json = String();
-    if (!json.reserve(measuredLength))
-    {
-        debugW("Unable to reserve %zu bytes for setting specs JSON.", measuredLength);
-        return false;
-    }
-
-    const size_t serializedLength = serializeJson(jsonArray, json);
-    if (serializedLength != measuredLength)
-    {
-        debugW("Setting specs JSON length mismatch: measured=%zu serialized=%zu.", measuredLength, serializedLength);
-        return false;
-    }
-
-    return true;
-}
-
-bool CWebServer::EnsureDeviceSettingSpecsJson()
-{
-    std::lock_guard guard(g_settingSpecsCacheMutex);
-
-    if (deviceSettingSpecsJson.length() > 0)
-        return true;
-
-    // Build into a local String first. deviceSettingSpecsJson is used as a
-    // write-once response backing buffer, so after this successful assignment
-    // AsyncWebServer responses can safely reference its c_str() storage.
-    String json;
-    if (!BuildSettingSpecsJson(json, LoadDeviceSettingSpecs()))
-    {
-        return false;
-    }
-
-    deviceSettingSpecsJson = json;
-    debugI("WebServer: cached device setting specs JSON (%zu bytes).", (size_t)deviceSettingSpecsJson.length());
-    return true;
-}
-
-const std::vector<std::reference_wrapper<SettingSpec>> & CWebServer::LoadDeviceSettingSpecs()
-{
-    std::lock_guard guard(g_settingSpecsCacheMutex);
-
-    if (deviceSettingSpecs.empty())
-    {
-        // effectInterval lives on the EffectManager rather than DeviceConfig,
-        // so its spec is owned here. The Widget metadata mirrors the legacy
-        // composite "Rotate effects toggle + seconds input" UX in a way the
-        // front-end can render generically.
-        mySettingSpecs.push_back(SettingSpec::Validate(SettingSpec{
-            .Name                = "effectInterval",
-            .FriendlyName        = "Effect interval",
-            .Description         = "The duration in milliseconds that an individual effect runs, before the next effect is activated. "
-                                   "Disable rotation to keep the current effect active indefinitely.",
-            .Type                = SettingSpec::SettingType::PositiveBigInteger,
-            .Section             = "appearance",
-            .ApiPath             = "effects.effectInterval",
-            .Widget              = SettingSpec::WidgetKind::IntervalToggle,
-            .IntervalUnitDivisor = 1000,
-            .IntervalOnLabel     = "Rotate effects",
-            .IntervalOffLabel    = "Off",
-            .IntervalUnitLabel   = "seconds"
-        }));
-
-        deviceSettingSpecs.insert(deviceSettingSpecs.end(), mySettingSpecs.begin(), mySettingSpecs.end());
-
-        auto deviceConfigSpecs = g_ptrSystem->GetDeviceConfig().GetSettingSpecs();
-        deviceSettingSpecs.insert(deviceSettingSpecs.end(), deviceConfigSpecs.begin(), deviceConfigSpecs.end());
-    }
-
-    return deviceSettingSpecs;
-}
-
-void CWebServer::GetSettingSpecs(AsyncWebServerRequest * pRequest)
-{
-    // This is a high-traffic endpoint since the front-end fetches it on every page load
-    // to dynamically render the settings UI, so we cache the serialized JSON rather than
-    // re-serializing on every request. If the cache is empty for any reason (e.g. JSON
-    // document overflow during cache population), we attempt to rebuild it on demand and
-    // respond with an error if that fails.
-
-    if (!EnsureDeviceSettingSpecsJson())
-    {
-        SendBufferOverflowResponse(pRequest);
-        return;
-    }
-
-    auto response = pRequest->beginResponse(HttpOk, "application/json",
-        reinterpret_cast<const uint8_t*>(deviceSettingSpecsJson.c_str()),
-        deviceSettingSpecsJson.length());
-    AddCORSHeaderAndSendResponse(pRequest, response);
-}
-
-// Responds with current config, excluding any sensitive values
-void CWebServer::GetSettings(AsyncWebServerRequest * pRequest)
-{
-    debugV("GetSettings");
-
-    auto response = new AsyncJsonResponse();
-    response->addHeader("Server", "NightDriverStrip");
-    auto root = response->getRoot();
-    JsonObject jsonObject = root.to<JsonObject>();
-
-    // We get the serialized JSON for the device config, without any sensitive values
-    g_ptrSystem->GetDeviceConfig().SerializeToJSON(jsonObject, false);
-    jsonObject["effectInterval"] = g_ptrSystem->GetEffectManager().GetInterval();
-
-    AddCORSHeaderAndSendResponse(pRequest, response);
-}
-
-void CWebServer::GetUnifiedSettings(AsyncWebServerRequest * pRequest)
-{
-    auto response = new AsyncJsonResponse();
-    auto root = response->getRoot().to<JsonObject>();
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    deviceConfig.SerializeUnifiedSettings(root);
-    root["effects"]["effectInterval"] = g_ptrSystem->GetEffectManager().GetInterval();
-    AddCORSHeaderAndSendResponse(pRequest, response);
-}
-
-void CWebServer::GetUnifiedSettingsSchema(AsyncWebServerRequest * pRequest)
-{
-    auto response = new AsyncJsonResponse();
-    auto root = response->getRoot().to<JsonObject>();
-    g_ptrSystem->GetDeviceConfig().SerializeUnifiedSettingsSchema(root);
-    AddCORSHeaderAndSendResponse(pRequest, response);
-}
-
-bool CWebServer::ValidateLegacyDeviceSettings(AsyncWebServerRequest * pRequest, String* errorMessage)
-{
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-
-    // Validate the constrained settings first so the legacy POST path behaves like the unified JSON API:
-    // either the nontrivial request is coherent as a whole, or we reject it before any setters persist a
-    // partially applied config.
-
-    if (pRequest->hasParam(DeviceConfig::OpenWeatherApiKeyTag, true, false))
-    {
-        auto [isValid, validationMessage] =
-            deviceConfig.ValidateOpenWeatherAPIKey(pRequest->getParam(DeviceConfig::OpenWeatherApiKeyTag, true, false)->value());
-        if (!isValid)
-        {
-            if (errorMessage)
-                *errorMessage = validationMessage;
-            return false;
-        }
-    }
-
-    if (pRequest->hasParam(DeviceConfig::PowerLimitTag, true, false))
-    {
-        auto [isValid, validationMessage] =
-            deviceConfig.ValidatePowerLimit(pRequest->getParam(DeviceConfig::PowerLimitTag, true, false)->value().toInt());
-        if (!isValid)
-        {
-            if (errorMessage)
-                *errorMessage = validationMessage;
-            return false;
-        }
-    }
-
-    if (pRequest->hasParam(DeviceConfig::BrightnessTag, true, false))
-    {
-        auto [isValid, validationMessage] =
-            deviceConfig.ValidateBrightness(pRequest->getParam(DeviceConfig::BrightnessTag, true, false)->value().toInt());
-        if (!isValid)
-        {
-            if (errorMessage)
-                *errorMessage = validationMessage;
-            return false;
-        }
-    }
-
-    if (pRequest->hasParam(DeviceConfig::AudioInputPinTag, true, false))
-    {
-        auto [isValid, validationMessage] =
-            deviceConfig.ValidateAudioInputPin(pRequest->getParam(DeviceConfig::AudioInputPinTag, true, false)->value().toInt());
-        if (!isValid)
-        {
-            if (errorMessage)
-                *errorMessage = validationMessage;
-            return false;
-        }
-    }
-
-    if (errorMessage)
-        *errorMessage = "";
-
-    return true;
-}
-
-// Support function that applies whatever settings are included in the request passed after validating the
-// nontrivial ones up front. Returning false lets the caller report a single coherent bad-request error
-// instead of half-applying device config and then discovering a later constraint violation.
-bool CWebServer::SetSettingsIfPresent(AsyncWebServerRequest * pRequest, String* errorMessage)
-{
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    auto& effectManager = g_ptrSystem->GetEffectManager();
-
-    if (!ValidateLegacyDeviceSettings(pRequest, errorMessage))
-        return false;
-
-    auto runtimeConfig = deviceConfig.GetRuntimeConfig();
-    bool runtimeConfigChanged = false;
-
-    runtimeConfigChanged = PushPostParamIfPresent<size_t>(pRequest, DeviceConfig::MatrixWidthTag, SET_VALUE(runtimeConfig.topology.width = value)) || runtimeConfigChanged;
-    runtimeConfigChanged = PushPostParamIfPresent<size_t>(pRequest, DeviceConfig::MatrixHeightTag, SET_VALUE(runtimeConfig.topology.height = value)) || runtimeConfigChanged;
-    runtimeConfigChanged = PushPostParamIfPresent<bool>(pRequest, DeviceConfig::MatrixSerpentineTag, SET_VALUE(runtimeConfig.topology.serpentine = value)) || runtimeConfigChanged;
-    runtimeConfigChanged = PushPostParamIfPresent<size_t>(pRequest, DeviceConfig::WS281xChannelCountTag, SET_VALUE(runtimeConfig.outputs.channelCount = value)) || runtimeConfigChanged;
-    runtimeConfigChanged = PushPostParamIfPresent<String>(pRequest, DeviceConfig::OutputDriverTag, SET_VALUE(runtimeConfig.outputs.driver = value == "hub75" ? DeviceConfig::OutputDriver::HUB75 : DeviceConfig::OutputDriver::WS281x)) || runtimeConfigChanged;
-    runtimeConfigChanged = PushPostParamIfPresent<String>(pRequest, DeviceConfig::WS281xColorOrderTag, SET_VALUE(
-        if (value == "RGB") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::RGB;
-        else if (value == "RBG") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::RBG;
-        else if (value == "GRB") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::GRB;
-        else if (value == "GBR") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::GBR;
-        else if (value == "BRG") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::BRG;
-        else if (value == "BGR") runtimeConfig.outputs.colorOrder = DeviceConfig::WS281xColorOrder::BGR;
-    )) || runtimeConfigChanged;
-
-    if (runtimeConfigChanged)
-    {
-        auto [isValid, validationMessage] = deviceConfig.ValidateRuntimeConfig(runtimeConfig);
-        if (!isValid)
-        {
-            if (errorMessage)
-                *errorMessage = validationMessage;
-            return false;
-        }
-    }
-
-    if (runtimeConfigChanged)
-    {
-        String runtimeErrorMessage;
-        if (!g_ptrSystem->ApplyRuntimeConfigurationTransaction(runtimeConfig, &runtimeErrorMessage))
-        {
-            if (errorMessage)
-                *errorMessage = runtimeErrorMessage;
-            return false;
-        }
-    }
-
-    PushPostParamIfPresent<size_t>(pRequest,"effectInterval", SET_VALUE(effectManager.SetInterval(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::HostnameTag, SET_VALUE(deviceConfig.SetHostname(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::LocationTag, SET_VALUE(deviceConfig.SetLocation(value)));
-    PushPostParamIfPresent<bool>(pRequest, DeviceConfig::LocationIsZipTag, SET_VALUE(deviceConfig.SetLocationIsZip(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::CountryCodeTag, SET_VALUE(deviceConfig.SetCountryCode(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::OpenWeatherApiKeyTag, SET_VALUE(deviceConfig.SetOpenWeatherAPIKey(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::TimeZoneTag, SET_VALUE(deviceConfig.SetTimeZone(value)));
-    PushPostParamIfPresent<bool>(pRequest, DeviceConfig::Use24HourClockTag, SET_VALUE(deviceConfig.Set24HourClock(value)));
-    PushPostParamIfPresent<bool>(pRequest, DeviceConfig::UseCelsiusTag, SET_VALUE(deviceConfig.SetUseCelsius(value)));
-    PushPostParamIfPresent<String>(pRequest, DeviceConfig::NTPServerTag, SET_VALUE(deviceConfig.SetNTPServer(value)));
-    PushPostParamIfPresent<bool>(pRequest, DeviceConfig::RememberCurrentEffectTag, SET_VALUE(deviceConfig.SetRememberCurrentEffect(value)));
-    PushPostParamIfPresent<int>(pRequest, DeviceConfig::PowerLimitTag, SET_VALUE(deviceConfig.SetPowerLimit(value)));
-    PushPostParamIfPresent<int>(pRequest, DeviceConfig::BrightnessTag, SET_VALUE(deviceConfig.SetBrightness(value)));
-
-    // Audio input pin: persist the new pin and, when supported, ask AudioService
-    // to apply it live so the user doesn't need to reboot.
-    {
-        const int oldPin = deviceConfig.GetAudioInputPin();
-        const bool audioInputPinChanged =
-            PushPostParamIfPresent<int>(pRequest, DeviceConfig::AudioInputPinTag, SET_VALUE(deviceConfig.SetAudioInputPin(value)));
-        if (audioInputPinChanged && !ApplyAudioInputPinChange(oldPin))
-        {
-            if (errorMessage)
-                *errorMessage = "Audio input pin live apply failed";
-            return false;
-        }
-    }
-
-    #if SHOW_VU_METER
-    PushPostParamIfPresent<bool>(pRequest, DeviceConfig::ShowVUMeterTag, SET_VALUE(effectManager.ShowVU(value)));
-    #endif
-
-    std::optional<CRGB> globalColor = {};
-    std::optional<CRGB> secondColor = {};
-
-    PushPostParamIfPresent<CRGB>(pRequest, DeviceConfig::GlobalColorTag, SET_VALUE(globalColor = value));
-    PushPostParamIfPresent<CRGB>(pRequest, DeviceConfig::SecondColorTag, SET_VALUE(secondColor = value));
-
-    deviceConfig.ApplyColorSettings(globalColor, secondColor,
-                                    IsPostParamTrue(pRequest, DeviceConfig::ClearGlobalColorTag),
-                                    IsPostParamTrue(pRequest, DeviceConfig::ApplyGlobalColorsTag));
-
-    if (errorMessage)
-        *errorMessage = "";
-
-    return true;
-}
-
-// Set settings and return resulting config
-void CWebServer::SetSettings(AsyncWebServerRequest * pRequest)
-{
-    debugV("SetSettings");
-
-    String errorMessage;
-    if (!SetSettingsIfPresent(pRequest, &errorMessage))
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, errorMessage);
-        return;
-    }
-
-    // We return the current config in response
-    GetSettings(pRequest);
-}
-
-void CWebServer::SetUnifiedSettings(AsyncWebServerRequest * pRequest, JsonVariantConst json)
-{
-    if (!json.is<JsonObjectConst>())
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, "Malformed request");
-        return;
-    }
-
-    auto root = json.as<JsonObjectConst>();
-    auto& deviceConfig = g_ptrSystem->GetDeviceConfig();
-    DeviceConfig::UnifiedSettingsRequest unifiedRequest;
-
-    auto [isValid, validationMessage] = deviceConfig.ParseAndValidateUnifiedSettings(root, unifiedRequest);
-    if (!isValid)
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, validationMessage);
-        return;
-    }
-
-    if (unifiedRequest.runtimeConfigTouched)
-    {
-        String runtimeErrorMessage;
-        if (!g_ptrSystem->ApplyRuntimeConfigurationTransaction(unifiedRequest.requestedRuntimeConfig, &runtimeErrorMessage))
-        {
-            AddCORSHeaderAndSendBadRequest(pRequest, runtimeErrorMessage);
-            return;
-        }
-    }
-
-    String applyErrorMessage;
-    if (!deviceConfig.ApplyUnifiedDeviceSettings(unifiedRequest, &applyErrorMessage))
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, applyErrorMessage);
-        return;
-    }
-
-    auto& effectManager = g_ptrSystem->GetEffectManager();
-
-    if (root["effects"].is<JsonObjectConst>())
-    {
-        auto effects = root["effects"].as<JsonObjectConst>();
-        if (effects["effectInterval"].is<size_t>())
-            effectManager.SetInterval(effects["effectInterval"].as<size_t>());
-    }
-
-    GetUnifiedSettings(pRequest);
-}
-
-bool CWebServer::CheckAndGetSettingsEffect(AsyncWebServerRequest * pRequest, std::shared_ptr<LEDStripEffect> & effect, bool post)
-{
-    auto effectIndex = GetEffectIndexFromParam(pRequest, post);
-
-    if (effectIndex < 0)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-
-        return false;
-    }
-
-    effect = g_ptrSystem->GetEffectManager().EffectAt(effectIndex);
-    if (!effect)
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return false;
-    }
-
-    return true;
-}
-
-void CWebServer::GetEffectSettingSpecs(AsyncWebServerRequest * pRequest)
-{
-    std::shared_ptr<LEDStripEffect> effect;
-
-    if (!CheckAndGetSettingsEffect(pRequest, effect))
-        return;
-
-    std::vector<std::reference_wrapper<SettingSpec>> settingSpecs;
-    {
-        std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-        settingSpecs = effect->GetSettingSpecs();
-    }
-
-    SendSettingSpecsResponse(pRequest, settingSpecs);
-}
-
-void CWebServer::SendEffectSettingsResponse(AsyncWebServerRequest * pRequest, std::shared_ptr<LEDStripEffect> & effect)
-{
-    auto response = std::make_unique<AsyncJsonResponse>();
-    auto jsonObject = response->getRoot().to<JsonObject>();
-
-    bool serialized = false;
-    {
-        std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-        serialized = effect->SerializeSettingsToJSON(jsonObject);
-    }
-
-    if (serialized)
-    {
-        AddCORSHeaderAndSendResponse(pRequest, response.release());
-        return;
-    }
-
-    debugV("JSON response buffer overflow!");
-    SendBufferOverflowResponse(pRequest);
-}
-
-void CWebServer::GetEffectSettings(AsyncWebServerRequest * pRequest)
-{
-    debugV("GetEffectSettings");
-
-    std::shared_ptr<LEDStripEffect> effect;
-
-    if (!CheckAndGetSettingsEffect(pRequest, effect))
-        return;
-
-    SendEffectSettingsResponse(pRequest, effect);
-}
-
-bool CWebServer::ApplyEffectSettings(AsyncWebServerRequest * pRequest, std::shared_ptr<LEDStripEffect> & effect)
-{
-    bool settingChanged = false;
-
-    // Effect settings are live data consumed by Draw(). Apply and enumerate
-    // them under the render/effect locks so web requests cannot mutate an
-    // effect while the render task is drawing it.
-    std::scoped_lock guard(g_render_mutex, g_effect_manager_mutex);
-    for (auto& settingSpecWrapper : effect->GetSettingSpecs())
-    {
-        const auto& spec = settingSpecWrapper.get();
-        // Settings with an ApiPath are addressed via a structured path, not by name — skip them here.
-        if (spec.ApiPath)
-            continue;
-        const String& settingName = spec.Name;
-        settingChanged = PushPostParamIfPresent<String>(pRequest, settingName, [&](auto value) { return effect->SetSetting(settingName, value); })
-            || settingChanged;
-    }
-
-    return settingChanged;
-}
-
-void CWebServer::SetEffectSettings(AsyncWebServerRequest * pRequest)
-{
-    debugV("SetEffectSettings");
-
-    std::shared_ptr<LEDStripEffect> effect;
-
-    if (!CheckAndGetSettingsEffect(pRequest, effect, true))
-        return;
-
-    if (ApplyEffectSettings(pRequest, effect))
-        SaveEffectManagerConfig();
-
-    SendEffectSettingsResponse(pRequest, effect);
-}
-
-// Validate and set one setting. If no validator is available in settingValidators for the setting, validation is skipped.
-//   Requests containing more than one known setting are malformed and rejected.
-void CWebServer::ValidateAndSetSetting(AsyncWebServerRequest * pRequest)
-{
-    String paramName;
-
-    for (auto& settingSpecWrapper : LoadDeviceSettingSpecs())
-    {
-        auto& settingSpec = settingSpecWrapper.get();
-
-        // Settings with an ApiPath are addressed via a structured path, not by name — skip them here.
-        if (settingSpec.ApiPath)
-            continue;
-
-        if (pRequest->hasParam(settingSpec.Name, true))
-        {
-            if (paramName.isEmpty())
-                paramName = settingSpec.Name;
-            else
-            // We found multiple known settings in the request, which we don't allow
-            {
-                AddCORSHeaderAndSendBadRequest(pRequest, "Malformed request");
-                return;
-            }
-        }
-    }
-
-    // No known setting in the request, so we can stop processing and go on with our business
-    if (paramName.isEmpty())
-    {
-        AddCORSHeaderAndSendOKResponse(pRequest);
-        return;
-    }
-
-    auto validator = settingValidators.find(paramName);
-    if (validator != settingValidators.end())
-    {
-        const String &paramValue = pRequest->getParam(paramName, true)->value();
-        bool isValid;
-        String validationMessage;
-
-        std::tie(isValid, validationMessage) = validator->second(paramValue);
-
-        if (!isValid)
-        {
-            AddCORSHeaderAndSendBadRequest(pRequest, validationMessage);
-            return;
-        }
-    }
-
-    // Process the setting as per usual
-    String errorMessage;
-    if (!SetSettingsIfPresent(pRequest, &errorMessage))
-    {
-        AddCORSHeaderAndSendBadRequest(pRequest, errorMessage);
-        return;
-    }
-    AddCORSHeaderAndSendOKResponse(pRequest);
-}
 
 // Reset effect config, device config and/or the board itself
 void CWebServer::Reset(AsyncWebServerRequest * pRequest)
